# preference optimization sequence length metric logging fix