--- a/src/fairseq2/recipes/lm/preference_finetune/utils.py
+++ b/src/fairseq2/recipes/lm/preference_finetune/utils.py
@@ -144,10 +144,10 @@ class PreferenceFinetuneMetricBag(SequenceMetricBag):
             The batch processed by the model.
         """
         self._chosen_lengths.update(
-            Tensor(batch.chosen.num_target_elements()) / batch.chosen.batch_size,
+            Tensor([batch.chosen.num_target_elements() / batch.chosen.batch_size]),
             weight=batch.chosen.batch_size,
         )
         self._rejected_lengths.update(
-            Tensor(batch.rejected.num_target_elements()) / batch.rejected.batch_size,
+            Tensor([batch.rejected.num_target_elements() / batch.rejected.batch_size]),
             weight=batch.rejected.batch_size,
         )