**What does this PR do? Please describe:**
Fixes over/under flow Tensor bug for sequence length logging with preference optimization

Output before: 
```
[Rank 0] 2024-08-26 20:16:06,765 INFO fairseq2.recipes.trainer - Train Metrics (step 1) - DPO Loss: 7.71594 | Chosen Sequence Log Probabilities: -507.598 | Rejected Sequence Log Probabilities: -408.038 | Chosen Sequence Length: -3.4783e+32 | Rejected Sequence Length: nan | NLL Loss: 1.89538 | Gradient Norm: 1801.15 | Elapsed Time: 14s | Wall Time: 55s | Learning Rate: 5.49998e-06 | Batch Size: 52 | Elements per Batch: 16,564 | Elements per Second: 4,615 | Number of Examples: 209 | Number of Elements: 66,257 | Number of Target Elements: 56,008 | Total Number of Examples: 209 | Total Number of Elements: 66,257 | Total Number of Target Elements: 56,008
```

Output now:
```
[Rank 0] 2024-08-28 02:11:37,034 INFO fairseq2.recipes.trainer - Train Metrics (step 1) - DPO Loss: 7.71594 | Chosen Sequence Log Probabilities: -507.598 | Rejected Sequence Log Probabilities: -408.038 | Chosen Sequence Length: 267.981 | Rejected Sequence Length: 203.651 | NLL Loss: 1.89538 | Gradient Norm: 1813.48 | Elapsed Time: 14s | Wall Time: 55s | Learning Rate: 5.49998e-06 | Batch Size: 52 | Elements per Batch: 16,564 | Elements per Second: 4,729 | Number of Examples: 209 | Number of Elements: 66,257 | Number of Target Elements: 56,008 | Total Number of Examples: 209 | Total Number of Elements: 66,257 | Total Number of Target Elements: 56,008
```

**Does your PR introduce any breaking changes? If yes, please list them:**
None

**Check list:**
- [ ] Was the content of this PR **discussed and approved** via a GitHub issue? (no need for typos or documentation improvements)
- [ ] Did you read the [contributor guideline](https://github.com/facebookresearch/fairseq2/blob/main/CONTRIBUTING.md)?
- [ ] Did you make sure that your **PR does only one thing** instead of bundling different changes together?
- [ ] Did you make sure to **update the documentation** with your changes? (if necessary)
- [ ] Did you write any **new necessary tests**?
- [ ] Did you verify new and **existing tests pass** locally with your changes?
- [ ] Did you **update the [CHANGELOG](https://github.com/facebookresearch/fairseq2/blob/main/CHANGELOG.md)**? (no need for typos, documentation, or minor internal changes)